refactor(api, chat): extract notebook stream core and add configurable endpoint
Extract shared SSE streaming logic into notebookStreamCore.ts for reuse
by authenticated and public endpoints. Add configurable endpoint prop to
NotebookChatProvider and 429 rate limit error handling to the adapter.
